requirements.txt
primer3-py
pyfaidx